refactor: extract session title actions from AgentWorkspace
